refactor data storage to provider pattern
